+++ graphql-java-runtime/src/main/java/com/graphql_java_generator/client/QueryExecutorImpl.java
@@ -39,6 +39,7 @@ import com.graphql_java_generator.exception.GraphQLRequestExecutionException;
  * 
  * @author etienne-sf
  */
+@Deprecated
 public class QueryExecutorImpl implements QueryExecutor {
 
 	static {