+++ graphql-java-spring-mvc/src/main/java/graphql/spring/web/servlet/ExecutionResultHandler.java
@@ -1,0 +1,12 @@
+package graphql.spring.web.servlet;
+
+import graphql.ExecutionResult;
+import graphql.PublicSpi;
+
+import java.util.concurrent.CompletableFuture;
+
+@PublicSpi
+public interface ExecutionResultHandler {
+
+    Object handleExecutionResult(CompletableFuture<ExecutionResult> executionResultCF);
+}